Remove MockProfileViewModel and its related engineering code
