+++ app/src/debug/AndroidManifest.xml
@@ -67,16 +67,6 @@
             android:theme="@style/Hedvig.Theme" />
 
         <activity
-            android:name=".feature.profile.ProfileMockActivity"
-            android:label="Home Mock Screen"
-            android:theme="@style/Hedvig.Theme" />
-
-        <activity
-            android:name=".feature.marketpicker.MarketPickerMockActivity"
-            android:label="Home Mock Screen"
-            android:theme="@style/Hedvig.Theme" />
-
-        <activity
             android:name=".feature.loggedin.LoggedInMockActivity"
             android:label="Logged In Mock Screen"
             android:theme="@style/Hedvig.Theme" />
@@ -94,8 +84,6 @@
         <activity
             android:name=".feature.checkout.CheckoutMockActivity"
             android:label="Checkout Mock Screen" />
-
-        <activity android:name=".feature.payment.PaymentMockActivity" />
 
         <activity android:name=".feature.changeaddress.ChangeAddressMockActivity" />
