+++ app/src/engineering/java/com/hedvig/app/DevelopmentActivity.kt
@@ -18,7 +18,6 @@ import com.hedvig.app.feature.loggedin.LoggedInMockActivity
 import com.hedvig.app.feature.loggedin.ui.LoggedInActivity
 import com.hedvig.app.feature.offer.OfferMockActivity
 import com.hedvig.app.feature.onboarding.OnboardingMockActivity
-import com.hedvig.app.feature.profile.ProfileMockActivity
 import com.hedvig.app.feature.referrals.ReferralsMockActivity
 import com.hedvig.app.feature.sunsetting.ForceUpgradeActivity
 import com.hedvig.app.feature.tracking.TrackingLogActivity
@@ -79,9 +78,6 @@ class DevelopmentActivity : AppCompatActivity(R.layout.activity_development) {
                     },
                     DevelopmentScreenAdapter.DevelopmentScreenItem.Row("Insurance") {
                         startActivity(Intent(this, InsuranceMockActivity::class.java))
-                    },
-                    DevelopmentScreenAdapter.DevelopmentScreenItem.Row("Profile") {
-                        startActivity(Intent(this, ProfileMockActivity::class.java))
                     },
                     DevelopmentScreenAdapter.DevelopmentScreenItem.Row("Logged in") {
                         startActivity(Intent(this, LoggedInMockActivity::class.java))
