+++ app/src/main/java/com/hedvig/app/util/extensions/MutableLiveDataExtensions.kt
@@ -2,11 +2,6 @@ package com.hedvig.app.util.extensions
 
 import androidx.lifecycle.MutableLiveData
 
-fun <T> MutableLiveData<T>.default(initalValue: T): MutableLiveData<T> {
-    apply { setValue(initalValue) }
-    return this
-}
-
 /**
  * Takes current value and updates it with the result of [function] if current value is not null
  */
